Change __unicode__() calls to __str__()
Model instances were not showing human-readable string names, but were
instead displayed as "{model name} object."  The __str__() method is
recommended for python 3.

This commit also changes the direction and street type abbreviations
to be capitalized by altering their entries in the schema.  @xray7224,
is that the right way to capitalize these in Django?  I wanted an
address to display as 1 N Streetname Rd instead of 1 n Streetname rd.