Add map object.
Add an app (django_google_maps) that inserts a Google Map on the
change form for a given object.  This commit adds the google map only
to a test object (mapTest), but the goal is to include a map on the
events and participants change forms (commits forthcoming).

This commit also includes a slew of migrations, some of which are
overdue from past testing changes.  Migrations 0009_ and beyond all
have to do with adding the maps.

In this commit I added a new model (mapTest), but it will go away in
future commits.  It was purely for testing the map embedding.