#6: Add map to event change form.
This commit adds a map to the event change form and sets the map to be drawn on
page load, from the existing address, instead of on address change.  It also
removes the mapTest object, since I only added that as a test before adding the
maps to the participants and events models.