#14: Add terminal event to event model
Add optional link to another event as the "terminal" event for prep events.
That is, when users are organizing a big action, like a protest at City Hall,
they will have planning meetings leading up to that big action.  Each of those
planning meetings will be designated as a "prep" event.  The big action will be
referenced in the "terminal" field of each of those planning meetings.