#13: Use jquery to display participants on event page.
Change all functions to use jquery instead of straight Javascript (notably
append() instead of innerHTML, and .getJSON() instead of straight HTML GET).  To
do so, add jquery.js and include it in the template.  I removed @cwebber's
example table rows, but will continue to use them as examples from the history.

As of this commit, participants in a given event will display and the edit and
cancel buttons work.  Saving and removing participants does not yet work, nor
does adding a new person (working on it).