#13: Add new functions
Add savePerson(), getAvailableParticipants(), linkPerson() functions.  As of
this commit, savePerson() doesn't work and getAvailableParticipants() works but
is sloppy in how it inserts information into the page.  I haven't tested
linkPerson(), but modeled it on unlinkPerson().

I changed displayPerson() so it no longer uses a loop and matches up correctly
with the column headers we have.

This is still unfinished and doesn't yet use xPath to show/hide page elements as
it should.  I expect more from me and @cwebber tomorrow.