#13: Add new person, link existing person
With this commit, fetch available participants and link an existing person to
the event both work.  I added a line for adding a new person, too.  I'm going to
check with @cwebber tomorrow about formatting posts and puts properly so that I
can save changes to a person and add a new person via these forms.  I'm sure
we'll also do some reformatting.

I still have to make the edit and cancel buttons work properly, so that the
displayed and editable information about a person swap back and forth smoothly.