#13 insertParticipant function, and stubbed out friends
The design of this is done in such a way that rows can be cleared and
re-filled via the fill[Static,Edit,Errors]Row() methods.  So for
example, if a user "cancels" their current editing, we can just reset
the edit row back to a prior state by wiping out its elements and
replacing them with logical elements based on the json / object
representation we have of the participant.

Side note: we've got person, participant, attendee all as variable
names refering to the same thing.  It's confusing me, so I'm trying to
make new code say "participant".