#13: Update and save an existing participant
Just what it says on the tin!  As of now, we can change information about an
attendee from the sign-in-sheet.  The next commit will extend this to add a new
person altogether.
Note that at the moment there's an odd bug.  After editing and saving a
participant, the edit button stops being clickable.  Perhaps I have to
re-animate that callback as part of the recreateRows() function?