#13: Add institution, display new person as static row.
Add institution as a column on the sign-in sheet.  Make a newly created person
show up correctly as a static row on save.

The institution and address are currently not editable from the sign in sheet.
Both should be editable and optional (which they are not, currently).  Currently
participants created on the sign in sheet are not editable on the sign in sheet,
because they aren't linked to an institution.  This commit also includes a
workaround where new participants are linked to address "1" in order to be able
to save them without an address, which will break if another implementation
doesn't have an address with id "1."  The thing to do here is to show the
dropdowns of institutions and addresses as django does elsewhere.  Later I think
we want to make new institutions and addresses create-able from this page, but
probably not today.

This commit also removes the leftover ParticipantInline class from our failed
attempt at using built-in django functionality to make the sign-in-sheet form.

SaveParticipant() remains ugly for another commit.