#13: Fix empty institution bug
Prior to this commit, participant changes would not save if that participant was
not linked to an institution.  As @xray7224 suggested, I added a "continue" in
the case where the value was null.

This doesn't fix the problem in which a new institution can't be added from this
page.  That is an issue in inline_ajax.js.  I need to create an empty
participants.institution object if one does not already exist, then fill it with
the given name.