Add address migration.
Add a migration to set the address back to a foreign key field, as implied by
the changes in commit 2c23d64, whose message should have been:

Remove address-as-charfield.

This commit reverts the address to a Foreign Key field, as originally intended.
This means the address will not be editable on the sign-in-sheet form, but that
is being addressed in issue #29.