#31, update institution model.
Add contacts to the institution model.  This commit doesn't complete this task,
since we need to (1) limit the number of contacts to 4 correctly (this commit
includes a clean() function that starts this task, but doesn't work as
intended), (2) link to the contact's change forms, and (3) show contact phone
numbers in the inline.

These changes also relate to issue #21, since only admin users should be able to
designate contacts.