#27 and #29:  Add autocomplete app.
In many places, we want to convert the automatic dropdown of foreign keys and
many to many options into a text field with autocompletion.  The changes in this
commit don't accomplish this yet, but are on their way.  This commit requires
the autocomplete app in requirements.txt, but doesn't change the look of other
pages yet.