#34: use autocomplete for tags
Add autocompletion for tags on the institution profile.  Remove "tag" section
from the event profile and allow the default tag field (with autocompletion,
though it doesn't currently work).

The event page has a custom template which seems to conflict with the
autocomplete function.  We should add autocompletion to the participant first
and last name fields on the sign in sheet and on the "key contacts" section of
the institution profile, as well as to the even name dropdown for the "major
action" field.  These are complicated by the custom template and the inline
form.

Other places we may want to add autocompletion: the location field on event
page, the title field on key contacts, and potentially all fields on the new
participant screen and the sign in sheet.