#62 Allow user to edit the participant from event page twice
The previous solution did show the static row again, but accidentally "filled"
all the rows with the edit row.  The result of this was that the edit row
"became" the static row in the end... but since this was never supposed to
happen, the events that were previously fired by the edit button no longer
worked.