#47: Save rows that are mid-edit before submitting the event form
Add a new submit button that will replace the built-in "Done" button once I
finish the JQuery and move on to the CSS.  This new button looks for rows that
are mid-edit and saves them before submitting the entire page.  As of this
commit it works _except_ for the case where a new participant row has been added
and not yet saved.  One or more other participants will be saved if they are
being edited when the form is submitted.

Remaining: make this work for the "add new participant" case and make the CSS
changes to make this button look like the "Done" buttons on all other pages.  I
can probably organize some of this code into functions, too.