#47: Correct save with add new participant and update CSS
Saving the page by hitting "Done" now works even when adding a totally new
participant (via the "add new" button and an empty row).  I also updated the CSS
so this Done button, which I created for this purpose, mimics the standard Done
buttons across the codebase.  Naturally, I then hid the standard Done button so
that it won't appear on this page and confuse users.

#63 is still in full effect.  If you hit "Done" with an empty row, nothing will
 happen -- no saving, no errors.  I anticipate fixing that as part of #63.