#80 Fix table boundary misalignment
Wrap <legend> element in a containing <div> to allow managable positioning. <legend> isn't supposed to be wrapped in a <div>, but on the other hand <legend> is notoriously difficult to style (or we wouldn't even be having this issue). This fix allows the content in the <legend> to be positioned where we want it and also simplifies the css of the surrounding elements, i.e., the top border of the containing fieldset does not need to be erased, <legend> doesn't require top, left, and right borders, etc.