#91 No date now yields blank instead of "(None)"
On "Select Event to Change" page, events without dates are now displayed with a blank cell in the "date of action" column instead of "(None)". Unforunately this also had the effect of making the table unsortable by "date of action", but this problem was already an issue for the "attendee count" column as noted in #98.