#92 Fix missing calendar icon
It turned out that the broken calendar icon was not only a problem for Chrome. In fact it was broken on all browsers; it's just that nobody noticed because some browsers substituted clickable text (namely the word "calendar") for the missing icon instead of displaying the broken icon link that showed up in Chrome. This was happening because the DateTimeShortcuts.js library, which generates the link to the calendar icon, was using a Django variable called "admin_media_prefix" that is no longer used since it was deprecated in Django 1.4 (can you tell I've spent some time researhcing this?). The point is I needed to set this deprecated variable at a point before DateTimeShortcuts gets loaded and base_site.html seemed the best place to do that.