Fix #59 - Remove django_google_maps
This removes the parts of the migrations responsible for adding, altering
and removing the django_google_maps model and fields. I have carefully modified
the migrations and removed ones which overall make no net change in regard to
django_google_maps. I have also removed it from our settings and the specific
CSS for the maps.