#63: Sign-in sheet: show errors
When an invalid phone number or name is entered for a new or existing attendee
on the sign-in sheet, display an error to users.  Generally, this happens if the
phone number isn't valid (might not have an area code or be in an unacceptable
format), and/or the name is missing.

This "handleJSONErrors()" function can be expanded to include as many errors as
we need.  Once we handle adding institutions from the sign-in sheet, we will
also need to return errors in cases of invalidity there.