#90: Unlink contacts from institutions.
Add "remove" button to replace the remove checkboxes from the Key
Contacts area on the Institutions change form.  As of this commit, the
contact is removed in the backend but doesn't disappear from the page.
That will come next.  Also to come is adding the same "remove" button
when a new contact row is added to the page ("Add another Key Contact")
but not saved to the db.  Currently adding that new row makes a small
gray "x" appear in the remove column.  Naturally that column should
consistently show a remove button.

@xray7224, would you look at my API endpoint here if you get a chance?
I tried to follow what you did, and it is working, but I want to make
sure I'm not doing anything that's going to cause us problems later.
Since the contact is an intermediate many-to-many model, I just removed
the whole row from that table.  Both institution and participant remain,
but there's no record of their connection having existed.