#90: Hide contact row after removing
Even though commit be8c1c1 allowed users to remove a key contact from
the Institution page, the contact would remain visible until the page
was saved.  This is confusing behavior.  In this commit, I added a "hide
row" function to hide the contact from the page after it has been
removed from the institution in the database.