#90: Mimic "remove" button in new row
As noted in the issue, we wanted to replace the gray "x" with a "Remove"
button that looks the same as it does for the other rows.  I had some
trouble with this, since just replacing the "x" with a button doesn't
work correctly.  Clicking "Remove" would hide the empty row correctly,
but then prevented the user from opening another empty row.  Clicking
"Add another Key Contact" had no effect.

I'm not sure exactly how the previously-existing "x" worked.  I ended up
addressing this by styling the "x" so it looks (mostly) like the other
remove buttons.  The exact differences are that it doesn't change color
when hovered over and the background and font colors may be slightly
different.  I think it is also slightly smaller.  @nttaylor or another
CSS whiz can look at this and make it match more exactly if we like, but
it now functions as desired and mostly matches the other rows, as
specified in the issue.