#118: Add space between countdown and input
I've separated the countdown from the input and made it so the
right-justification of the text is constant. The tag_change.js file is
now a little longer than I'd like it to be.  In other words, I'd be
happy to see some code improvement here.  I separated the countdown
integer into its own span, and made that span have an absolute position
and width so that only the number would change, without changing the
position of the text to the right.