#117: Show empty string for nulls on change lists
Add the empty string as our "empty changelist value" for Institutions,
Actions, and Participants. I used [this Stack Overflow
response](http://stackoverflow.com/questions/28174881/in-the-django-admin-changelist-how-do-i-display-a-blank-space-instead-of-the-de)
to address this issue.  I reverted the change made in b95ad90 to fix
#91, since this solution will work for any model.  Theoretically we
could put this on all our models, but Institutions are unlikely to have
any null values (since only the Institution name is displayed on the
change list) and I'm not sure what behavior we want for the Log.  This
is better than our solution to #91 because it keeps the columns sortable
(and I think it is more Djangoic).