#58: Add title field for all Participants
Add a new title field for all Participants, separate from the one for
Key Contacts.  I suspect that in the future we will want to relate these
fields somehow.  It may be that we want to remove the contact-specific
"title" field and display the Participant.title field on the
Institutions page, in the list of Contacts.

Note that this migration includes a change to the verbose name of a
model.  I didn't realize that change would require a migration.  Since
it does, I should have made it separately, back in 21dc5c0.