#121: Remove Contact title
Now that we have a Participant title field, as of commit 3e82dd0, we can
remove the Contact-specific title and just show the Participant title in
its place.  This means that #74 will need to be expanded to include
adding and editing the title from the Contacts page, but I think that
once we have it working for either title or phone number it should be
simple to use the same system for both.