#108: Add "Save Changes" button to action fieldset
We had hidden the "Save and stay on page" button on all pages (see
commit d918a35) but I'm showing it on the Action page so that there is
an option to save the page without scrolling all the way past the
sign-in sheet.  I've displayed the button, moved it up into the first
fieldset (white box), and altered the styling to match the "Done"
button.