Use API to check saved state of models
This uses the API to check the state of the models that has been
saved. This allows us in future to handle the cases of specific
fieldsets which are saved by JS and can't be checked by comparing
the initial form data to the current.

The two problems with this approach are:
- window.onbeforeunload needs a string or not as a return value in a
  synchronous fashion but $.get(...); works asynchronously.

- How to get the API URL for the models and what to do for new models
  which don't have an explicitly defined API. This potentially could be
  fixed by adding a "expose_api" attribute to the model and if that is
  set having a new API view automatically created. The admin interface
  would then have to fill the API URL for each fieldset so the JS knows
  which API to call.