#94: Dynamically set fieldset header
Instead of including the header for the fieldset in the template file,
set it based on a field in the inline_form_config.  I added that field
to inline_form_config, removed the static header text and added an id to
that <h2> element.  Finally, I set the <h2> element's text to the value
of the new inline_form_config plural name field.