#94: Add profile links from Contacts/sign-in sheet
In acb6c45, @kfogel added links to the full profile of the Participant
mentioned in the Contact line.  I used the new inline_form_config setup
to add those links from both the Contact table and the sign-in sheet
form.  This is still mostly extensible because I've defined the URLs in
admin.py, but it does, unfortunately, test for the "name" column and
apply the link there.  There might be a nicer way to do this (always put
it in the first column?).