Fix Major Action toggling
The "toggle" was broken for whether an action is prep for a Major
Action.  This commit fixes it, though I want to look into the reason why
toggle has broken.  This commit does clean up the file quite a bit, but
the change shouldn't be necessary unless the jquery "toggle" command is
interacting with a class in an unexpected way.  Evidently toggle and
bootstrap sometimes interfere with one another.