Add link to Institution profile from sign-in sheet.
This is a bit of a roundabout solution that builds on my work in 69f4ad8
and d42f259.  I created the `createProfileLink()` and
`getExistingInlinedModelProfileUrl()` functions in those commits, and
extend them here to take an optional url path argument.  Generally, we
want to use strings provided in inline_form_config.  This commit is an
intermediate solution so that we can have a link to an institutional
profile while I learn how to create the string for the url path in a
more consistent way.  The difficulty comes here because the Institution
is not actually the inlined model on the sign-in sheet, so I'll need to
come up with some other way to refer to it.