Change am/pm to use radio buttons.
In b32bdb6, @xray7224 added the dropdowns for 12 hours and am/pm, as
requested in #28.  @kfogel suggested that we use radio buttons instead
of a select for am/pm, so this commit makes that change.  The css
changes are to position the radio buttons next to the time dropdown
instead of below it.