#130 Superficial fix for new "create new" button
Align the upper "create new" button (repositioned in 47903e2) with the
right border of the main table in change-list type pages, unless the
table is less than 75% of screen width, which would bring the button too
close to (or overlapping onto) the "Select X to Change" text. In that
case, button will remain right-aligned with 75% of screen width.

Positioning of visual elements remains broken at smaller screen widths
(see #96), with elements appearing in illogical positions or vanishing
entirely (as is the case with the upper "create new" button discussed
above). This behavior should be addressed in a future update.