#123 X is now pencil for autocomplete fields
I still can't quite believe I figured out a way to do this, and
especially without hacking python code, but here is an entirely CSS
solution to this issue courtesy of my sheer bloody-mindedness.

Apparently this issue will soon go away as noted in #94, but this fix
will live for eternity in the repository.