#46 Sticky header rows in place
This CSS-only solution works because the column widths of both header columns and body columns are defined in the CSS (using percentages, so table still adjusts to screen size) as opposed to having their widths determined by the browser based on the size of cell content. This allows us to align the columns btween table head and table body, which are now more disconnected than in a normal table.

"min-width" attributes on columns 1-3 are a kludgey hedge against these becoming too narrow such that a string of text with no whitespace (e.g., a person's first or last name) would render as wider than the width of its containing cell, which would lead to misalignment between header and body columns.