#46 use sticky headers on sign-up form
Expand use of sticky headers by broadening CSS rules for what they apply
to. Also tweak rules for counting child elements of a table row to
ensure that width rules are only applied to <td>s and <th>s. This allows
CSS to skip the hidden <input> element that is inserted into each table
row (which -IMO- doesn't belong there btw. This information would be
better transmitted via a data-* attribute in the <tr> element).

Columns have also all been made the same width to account for the
different kinds of information presented in different tables (e.g., one
table might have a person's name in the first column; another might have
an institution name in the first column, and so on.)