#46, #153: Remove fixed header from dynamic tables
Adding a floating/frozen header to the dynamically generated tables
caused some other troubles (see #153 for example).  It's also obscuring
the "New" that should show up next to the institution name on the
sign-in sheet.  We still want to make this change, but for now I think
it's causing some difficulties, so I'm reverting the first attempt.

@nttaylor, this was your commit so I just wanted to alert you that I'm
rolling it back for now.  We'll still have these changes in history, and
can work from them when we get the frozen/floating header working
properly on the js tables.