Fix #156 - Remove the is_staff requirement
Users do no have to have the is_staff flag set on them to
be able to login to the "admin" user interface now. This
didn't make sense as we're building the entire app on the admin
UI so every user should be able to access it unless they are
inactive (i.e. is_active is False).

This now hides the is_staff flag everywhere and removes any checks
that used it.