re-implement sticky headers #46
The CSS-only version had too many failings (misaligned column heads
after resizing screen to some dimensions, undesirable blank space under
tables with only a few rows, requirement to assign fixed widths to
columns, to name but three), so here is a much better javascript
solution. There is still (at least) one bug: the headers become unstuck
from their columns if the table is scrolled and then the browser window
is resized. (Resizing is fine as long as the table is in its default
position with the first row of data just under the header).