Fix #143: Reorder groups on tag change form
Note that the groups are *not* reordered in the other places they
appear.  I had hoped to find a way to do this globally, but the Meta
"ordering" option appears to be for ordering by a field, not by values
of a field.  Certainly there's a way to do it in SQL and tell Django
about that, but I'm not sure how.