Quick fix for #137
This commit does the easy fix for splitting address fields up into
different text fields.  It adds four db columns for the address parts
and applies them in a straightforward way, for both Participant and
Institution.  It also makes the "street address" input a little larger
than the others, since that data is likely to be longer as well.