Begin implement click-to-edit sign-up tables #135
User should be able to click in a table cell to edit the content of that
cell. Once this is working, the "edit" buttons in the right-hand column
can go away. Right now the switching works but autocomplete now fails on
those fields where it should happen. Newly "sticky" column headers also
break when table cells change size.