cont. implement click-to-edit sign-up tables #135
Click-to-edit now mostly works, but still not in the name field for some
reason. I've also only tested it on the Attendees page. But there at
least, clicking on a field makes it editable and clicking away from that
field saves the changes. Pressing "enter" also saves the changes if the
user is editing a given field.

This revision removes a lot of machinery that dealt with the separate
edit row that was part of the previous editing system, but there's still
more cleaning to do. The visual elements that are gone, however, are the
column of buttons that used to be in the far-right column of the table
and the "multi-action-submit" button (labeled "done") that used to be in
the lower-right corner of the page and was used to submit changes on all
open editing rows as well as the "upstairs" portion of the page. Since
the upstairs has its own submit button, no functionality should be lost
by eliminating the multi-action-submit button.