cont. implement click-to-edit sign-up tables #135
Re-add a rightmost column, but this time with just a circle with an "X"
in it that means "delete this row". This is the symmetrical
functionality to the "Add new" button.

The "delete row" buttons work. "Add new" is buggy as of this commit in
that hitting return/enter after typing a value in the first column
causes the entity to be added twice. Just moving focus out of the first
column only adds once, but information symbol and correct stylings do
not appear until page is refreshed.

This commit removes additional chunks of code that were written for the
old system of having a separate edit row for each tuple.