Update Participant API per #137
I removed the "address" column in commit 7f3039d.  It was still referred
to by the API, though.  This commit replaces it with the new street
address column.  Our current use of the API does not display or update
the address, but in future we may need to include the other address
columns as well.