Prep 135-click-to-edit for merge to master
Check value of "user_can_edit" from inline config to determine whether
user sees controls enabling him/her to add new or delete rows in
dynamically generated ajax tables. Currently, this value does not
reflect different privileges in different areas of the site (which it
evenutally should) but this will come in a future update.

Also made "document.ready" the first part of the script to execute
instead of setupInlinedModelCallbacks. Calls to check the inline config,
which as of this revision are necessary in setupInlinedModelCallbacks
were breaking the app because the inline config is not available until
after the document is ready.