put information "i" in circle #162
The profile link (which takes the form of the stylized letter "i" as
seen on tourist information kiosks) is now in a circle created entirely
by CSS. Lots of trial-and-error went into getting the "i" centered in
the circle, aligning it with the other text in the table cell, etc. Only
time and other users will tell if itlooks right on all browsers. I
minorly tweaked the structure around the delete buttons, which are also
circles made with CSS, as well to make them match the profile links in
construction.