font-size now constant in ajax tables #162
Our CSS was defining a larger font size for input elements than for
normal (read-only) divs in dynamically generated tables. This caused
table cells to shift jarringly when editing the contents of a cell.
Therefore the pertinent CSS rule has been removed. If larger text is
desirable, all text can be increased uniformly and according to the
user's preference by zooming in or out in the browser.