different permissions based on user & model #163
Low-level users (called "leaders", confusingly) can now add/edit/delete
attendees to events but cannot add/edit/delete contacts for
institutions.  In the meantime, higher-lever users can perform those
actions for both models.  Further differentiation and addition/removal
of various privileges for different users and different models should be
easier in the future as we now have client-side reference both to the
permissions of the logged-in user (a change implemented in admin.py) and
a way to determine which model (e.g., institution or event) is being
used (a change implemented in change_form.html).

This more differentiated approach means that the binary "user_can_edit"
value currently in a hidden input field can be retired.

Note that for low-level app-users, the "add new" button fails to appear
on the events change page (and it should appear).  The
appearance/non-appearance of this button is currently controlled from
inline_models.html and not inline_ajax.js, where the other changes were
made.  This will be corrected in a future update.  The logic that
decides whether or not the "add new" should appear should also be moved
to inline_ajax.js so that similar decisions about what controls are
available to a given user are all being made in the same place.