different permissions based on user & model #163
Tweak the way the decision is made as to whether the logged-in user can
edit the ajax tables.  We now use "user rank", which is based on the
user group name (e.g., "admin" outranks "staff") instead of looking at
which specific permissions are available to the logged-in user.  The
full set of permissions is no longer in the inline config.  However, if
we ever want client-side access to those permissions again, I've left
the requisite line in "admin.py" in commented-out form.

Note that this does not address the missing "add new" button. That
revision is still coming.