Link correct object, as in #167
This commit does only part of the work required by #167.  As of this
commit, the correct Participant is linked to the Action.  But, a new
Participant is _also_ created and linked to the Action.

I think removing the "make_editable" conditional in the
`linkInlinedModel()` function changed the behavior such that a new
static row, with the correct object, is added to the table of Attendees.
That static row is added the editable row is not closed, so that it
remains while the static row with the correct object appears beneath it.
This isn't the correct behavior, of course.  Now, even worse: I think
the row fetched in this line `saveInlinedModel($(this).closest("tr"));`
is not the correct row, which is why we're not finding the "id" in
`.data()` when we look for it in `saveInlinedModel()`.

@nttaylor, I think you probably have a better idea of how to test
whether we're passing the correct row to `saveInlinedModel()`.  I leave
it to you to decide how to handle the creation of the correct static row
-- maybe we're happy with that, and just need to hide the editable row
so that we don't ever ever risk creating/saving a duplicate (new)
object.  Let me know if this isn't clear.