Add preliminary TESTING.md
Now that we have more people working on the system, we need a way for
everyone to check their work to make sure that it isn't causing
regressions.  This came up first with the inlined models, so I've
started with the desired/expected behavior(s) there, but we should
expand it to cover other parts of the system, too.