add-new + autocomplete no longer adds to DB #167
The change made in 9cba607 seemed to not work in Chrome browsers.  This
commit introduces a hack which hides (i.e., sets css "visibility" to
"none") the row in which the user is operating.  Now when the "blur"
handler is called (which again, only happens as far as we know in
Chrome) the handler will first check its own row's visiblity.  If the
row is visible, the save function is called.  If not, it means that
autocomplete was used, so the save function is not called, because the
autocomplete handlers have taken care of all necessary operations.