Withhold certain info on Change User #175 #179
When doing a "change user" operation, display no longer shows
password-related gobbeldy-gook (#175) or boxes for cherry-picking single
permissions (#179; because permissions are determined by the user's
group).  This solution is substantively different from the previous
attempt to resolve #175, which resulted in other unwanted changes in
behavior.  This solution introduces control flow statements into the
Django template which define the specific conditions under which the
above content will be withheld.