Cast datetime to string to fix #184
When an archived participant was present on an attendee list, no
attendees appeared on the Action page.  The .serialize() method we're
calling on each participant can't serialize datetime objects, which led
to the error noted in #184.  This commit adds a check of the "archived"
value.  If it is non-null, then it casts it to a string.  This was
originally suggested by @nttaylor.  Thank you!