Create PhoneNumber class: #134
Remove phone number fields from Participant and replace with a link to
the new PhoneNumber class.  This is a very preliminary implementation
that doesn't meet our requirements by any means.  Next I'll need to add,
display, and edit multiple phone numbers for each participant.

I think this breaks the US-specific formatting, but that is a bit beside
the point since it also (I think) breaks the phone number display
everywhere.  I will fix all that in the next commit.  Hurray for
branches.