Show new phonenumber object in list_display: #134
I changed the US_display_phone function to show the latest phone number
added for the participant (theoretically).  However, in so doing I found
out that the phone number is somehow now being stored in US format,
instead of international format, so we're losing the part of the
function that checks whether the phone number is a US number or an
international one.

The phone number displays correctly in the participant change list, but
not in our ajaxy inlined tables.  Fixing that is for the next commit.