Cancel "add new" and clean up error row #177 #178
Add a check in the "blur" handler that removes a row if both 1) the
associated id is the empty string (meaning that it was just generated
with the "add new" button) and all of the fields are empty (meaning that
although the user clicked "add new" the user didn't enter any
information). This means that if the user clicks "add new" and then does
anything other than type into the first column (including but not
limited to clicking the "delete row" button), the new row will be
removed.

The remove operation takes out the associated error row (even though it
is never visible) along with the new client row.