Begin dev. of errors that appear in the table cell
Begin work on having validation errors for data entered via
click-to-edit appear directly in the table cell containing the invalid
entry. The table cell should also remain in "edit" mode to further
emphasize that the user needs to do more if they want the data to go
into the DB. Finally, the previous system involving a separate error row
with a message that indicated errors in the row below will be
removed. References to the separate error row (as well as the separate
editing row that was previously a part of the design) will be removed
from the code / renamed to reflect the new design.

Error descriptions also now originate properly from the server instead
of from inline_ajax.js.