Cont. dev. of errors that appear in the table cell
Add messages to console that show timing of event handlers vs. ajax
calls. This revision is not any closer to a solution; it only served to
help illustrate the conflict ineherent in a plan to put individualized
errors in relevant table cells in a system where all columns in a tuple
must contain valid values if any column is going to be updated.  The
consensus was to go forward with a plan to gray-out all other cells in a
table row (i.e., prevent the user from editing them) if any one table
cell contains an invalid value. Future commits will reflect this
decision.