Cont. dev. of errors that appear in the table cell
All functionality is now in place. If one column has an error, the cell
with the error gets a red border and the error message appears in the
errored cell. All other columns in that row become unavailable (which is
indicated by changing the background color to for editing. An alert box
appears if the user tries to edit another cell before correcting the
error. Correcting the error returns everything to normal.