Add institution to searching and change templates
This merges the searching and results page into one including
the advanced and basic search interface. It also adds the ability
to search based on an institution and have results which match on
participants who are contacts of that institution.

There is still some problems with regards to searching when a user
is not a participant in any Action even when no Action is specified.
This is due to the searching being largely using Action to start the
quries round.