Fix label-flipping problem #96
The template at swoptact/templates/admin/includes/fieldset.html features
a <div> with class "control-label" that surrounds a Django "{{
field.label_tag }}" selector. This was problematic. The "control-label"
class is properly only applied to a <label> element. When it is applied
to a <div> surrounding the <label>, as we had here, we get the
"flipping" problem described above.

However, the <label> tag is generated by Django and applying a custom
Bootstrap class to it would require a custom template tag. A less
invasive work-around was to apply a css "display: inline" rule to the
surrounding <div>.