Add new autocomplete widget to event fields: #78
Add persistent text widget to major_action and organizer fields on Event
model.  This has the undesired side effect of removing autocompletion on
Tags, because the tag autocompletion requires a different form to be
specified in admin.py.  I'm committing this to master at @cwebber's
recommendation, but it doesn't get us all the way to what I want.
Really, I want the persistent text widget on major_action and organizer
_and_ regular old autocompletion for tags.  This commit leaves tags with
the multiple-select that is standard for M2M fields.